+++ config.json
@@ -1,5 +1,5 @@
 {
-  "name": "Aluracord - Matrix (peas)",
+  "name": "Aluracord - Matrix",
   "theme": {
     "colors": {
       "primary": {
