+++ pages/chat.js
@@ -1,0 +1,3 @@
+export default function PaginaDoChat() {
+  return <div>Página do Chat</div>;
+}
